apps/api/pyproject.toml
[project]
name = "nra-api"
version = "0.1.0"
description = "Nat's Running App – FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "pydantic>=2.6",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.12",
    "python-jose[cryptography]>=3.3",
    "passlib[argon2]>=1.7",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "redis>=5.0",
    "rq>=1.15",
    "email-validator>=2.1",
    "ruff>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "types-python-jose",
    "types-redis",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

